requirements.txt
asyncpg==0.21.0
aiohttp==3.6.2
fastapi==0.61.1
pydantic[email]==1.6.1
typeguard==2.9.1
uvicorn==0.12.1
Shapely==1.7.1
mypy==0.782
pycodestyle==2.6.0
starlette-prometheus==0.7.0
nivacloud-logging==0.8.14
pytest==6.1.1
pytest-docker-compose==3.2.0
pytest-asyncio==0.14.0
python-dotenv==0.14.0
requests==2.24.0
gcloud-common-utils==0.2.0
httpx==0.16.0
respx==0.13.0
yapf==0.30.0
